pyproject.toml
[project]
name = "budgetrak"
version = "0.1.0"
description = "AI-powered budget tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.14.2",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.111.0",
    "google-generativeai>=0.8.3",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
budgetrak = "budgetrak.server:main"
